Add debug manager module and attach F1 window
