+++ CHANGELOG
@@ -1,0 +1,1 @@
+Add F1 Debug Window with live stats, actions, and variable tweaks
